Add files via upload
Task 1: Number Guessing Game
As part of my internship with CodSoft, I developed a Number Guessing Game in Java. The task involved creating a game where a random number between 1 and 100 is generated, and the user guesses the number while receiving feedback on whether the guess is correct, too high, or too low. The game continues until the correct number is guessed, with limits on the number of attempts and options for multiple rounds. The user's score can be tracked based on attempts or rounds won.